+++ src/components/TodoItem/TodoItem.tsx
@@ -1,0 +1,59 @@
+/* eslint-disable jsx-a11y/label-has-associated-control */
+/* eslint-disable jsx-a11y/control-has-associated-label */
+import React from 'react';
+import { Todo } from '../../types/Todo';
+import classNames from 'classnames';
+
+interface Props {
+  todo: Todo;
+  onDeleteTodos: (id: number) => void;
+  onHandleChecked: (id: number) => void;
+}
+
+export const TodoItem: React.FC<Props> = ({
+  todo,
+  onHandleChecked,
+  onDeleteTodos,
+}) => {
+  return (
+    <div
+      data-cy="Todo"
+      className={classNames('todo', { completed: todo.completed })}
+    >
+      <label className="todo__status-label">
+        <input
+          data-cy="TodoStatus"
+          type="checkbox"
+          className="todo__status"
+          checked={todo.completed}
+          onChange={() => onHandleChecked(todo.id)}
+        />
+      </label>
+
+      <span data-cy="TodoTitle" className="todo__title">
+        {todo.title}
+      </span>
+
+      {/* Remove button appears only on hover */}
+      <button
+        type="button"
+        className="todo__remove"
+        data-cy="TodoDelete"
+        onClick={() => onDeleteTodos(todo.id)}
+      >
+        ×
+      </button>
+
+      {/* overlay will cover the todo while it is being deleted or updated */}
+      <div
+        data-cy="TodoLoader"
+        className={classNames('modal overlay', {
+          // 'is-active': loading && selectedPost === todo.id,
+        })}
+      >
+        <div className="modal-background has-background-white-ter" />
+        <div className="loader" />
+      </div>
+    </div>
+  );
+};
